# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Central Studio & Stage Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOffer
      tags: [offers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOfferBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteOffer
      tags: [offers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}:
    get:
      operationId: getNotification
      tags: [notifications]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNotification
      tags: [notifications]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNotification
      tags: [notifications]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /marketing/campaigns:
    get:
      operationId: listCampaigns
      tags: [marketing]
      summary: List all marketing campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [marketing]
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /marketing/campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [marketing]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCampaign
      tags: [marketing]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCampaign
      tags: [marketing]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /marketing/campaigns/{id}/send:
    post:
      operationId: sendCampaign
      tags: [marketing]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendCampaignResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /package-orders:
    get:
      operationId: listPackageOrders
      tags: [packageOrders]
      summary: List all package orders
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: List of package orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackageOrder"
    post:
      operationId: createPackageOrder
      tags: [packageOrders]
      summary: Create a package order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageOrderBody"
      responses:
        "201":
          description: Created package order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageOrder"
  /package-orders/{id}:
    get:
      operationId: getPackageOrder
      tags: [packageOrders]
      summary: Get a package order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Package order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageOrder"
        "404":
          description: Not found
    patch:
      operationId: updatePackageOrder
      tags: [packageOrders]
      summary: Update a package order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackageOrderBody"
      responses:
        "200":
          description: Updated package order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageOrder"
    delete:
      operationId: deletePackageOrder
      tags: [packageOrders]
      summary: Delete a package order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Offer:
      type: object
      required: [id, title, discountPercent, isActive, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        discountPercent: { type: integer }
        validUntil: { type: ["string", "null"] }
        isActive: { type: boolean }
        classIds: { type: array, items: { type: integer } }
        createdAt: { type: string }
    CreateOfferBody:
      type: object
      required: [title, discountPercent]
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        discountPercent: { type: integer }
        validUntil: { type: ["string", "null"] }
        isActive: { type: boolean }
        classIds: { type: array, items: { type: integer } }
    UpdateOfferBody:
      type: object
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        discountPercent: { type: integer }
        validUntil: { type: ["string", "null"] }
        isActive: { type: boolean }
        classIds: { type: array, items: { type: integer } }
    Notification:
      type: object
      required: [id, title, body, target, isDraft, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        body: { type: string }
        target: { type: string }
        sentAt: { type: ["string", "null"] }
        isDraft: { type: boolean }
        createdAt: { type: string }
    CreateNotificationBody:
      type: object
      required: [title, body, target]
      properties:
        title: { type: string }
        body: { type: string }
        target: { type: string }
        isDraft: { type: boolean }
    UpdateNotificationBody:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
        target: { type: string }
        sentAt: { type: ["string", "null"] }
        isDraft: { type: boolean }
    Campaign:
      type: object
      required: [id, title, type, status, message, targetAudience, recipientCount, sentCount, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        type: { type: string }
        status: { type: string }
        subject: { type: ["string", "null"] }
        message: { type: string }
        targetAudience: { type: string }
        recipientCount: { type: integer }
        sentCount: { type: integer }
        scheduledAt: { type: ["string", "null"] }
        sentAt: { type: ["string", "null"] }
        createdAt: { type: string }
    CreateCampaignBody:
      type: object
      required: [title, type, message]
      properties:
        title: { type: string }
        type: { type: string }
        status: { type: string }
        subject: { type: ["string", "null"] }
        message: { type: string }
        targetAudience: { type: string }
        scheduledAt: { type: ["string", "null"] }
    UpdateCampaignBody:
      type: object
      properties:
        title: { type: string }
        type: { type: string }
        status: { type: string }
        subject: { type: ["string", "null"] }
        message: { type: string }
        targetAudience: { type: string }
        scheduledAt: { type: ["string", "null"] }
    SendCampaignResponse:
      type: object
      required: [success, sentCount]
      properties:
        success: { type: boolean }
        sentCount: { type: integer }
        sentAt: { type: ["string", "null"] }
...
    PackageOrder:
      type: object
      required: [id, studentName, studentEmail, packageName, totalCredits, remainingCredits, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        studentName: { type: string }
        studentEmail: { type: string }
        studentPhone: { type: ["string", "null"] }
        packageId: { type: ["integer", "null"] }
        packageName: { type: string }
        totalCredits: { type: integer }
        remainingCredits: { type: integer }
        status: { type: string }
        notes: { type: ["string", "null"] }
        activatedAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    CreatePackageOrderBody:
      type: object
      required: [studentName, studentEmail, packageName, totalCredits, remainingCredits]
      properties:
        studentName: { type: string }
        studentEmail: { type: string }
        studentPhone: { type: ["string", "null"] }
        packageId: { type: ["integer", "null"] }
        packageName: { type: string }
        totalCredits: { type: integer }
        remainingCredits: { type: integer }
        status: { type: string }
        notes: { type: ["string", "null"] }
        activatedAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
    UpdatePackageOrderBody:
      type: object
      properties:
        status: { type: string }
        remainingCredits: { type: integer }
        notes: { type: ["string", "null"] }
        activatedAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
...